app/requirements.txt
flask
prometheus_client
werkzeug
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
